requirements.txt
bibtexparser
requests
rich
citeproc-py
